Rewrote webpack settings
