--- environments/webpack/builders/favicon.js
@@ -1,19 +1,0 @@
-
-export const favicon = {
-  cache: false,
-  location: 'favicon',
-  original: '.favicon.png',
-  additionalVariants: [],
-};
-
-export const allVariants = [
-  'android',
-  'appleIcon',
-  'appleStartup',
-  'coast',
-  'firefox',
-  'opengraph',
-  'twitter',
-  'yandex',
-  'windows',
-];
