--- environments/webpack/builders/page.js
@@ -1,6 +1,0 @@
-
-export const page = {
-  title: 'Js Boilerplate',
-  template: '.index.ejs',
-  cache: false,
-};
