+++ environments/webpack/settings/index.js
@@ -1,7 +1,5 @@
-export * from './favicon-config';
+export * from './app';
 export * from './favicon';
-export * from './modules';
-export * from './page-config';
 export * from './page';
 export * from './paths';
 export * from './rules';
